Reformating and discovery unit tests.
