+++ app/src/main/java/com/speakeasy/watsonbarassistant/Discovery/OnTaskCompleted.kt
@@ -1,6 +1,6 @@
 package com.speakeasy.watsonbarassistant.Discovery
 
-import com.speakeasy.watsonbarassistant.DiscoveryRecipe
+import com.speakeasy.watsonbarassistant.Recipe.DiscoveryRecipe
 
 
 interface OnTaskCompleted {
