+++ app/src/main/java/com/speakeasy/watsonbarassistant/Discovery/SearchDiscovery.kt
@@ -4,8 +4,10 @@ import android.os.AsyncTask
 import android.util.Log
 import com.ibm.watson.developer_cloud.discovery.v1.Discovery
 import com.ibm.watson.developer_cloud.discovery.v1.model.QueryOptions
+import com.ibm.watson.developer_cloud.discovery.v1.model.QueryResponse
 import com.ibm.watson.developer_cloud.service.exception.NotFoundException
 import com.speakeasy.watsonbarassistant.*
+import com.speakeasy.watsonbarassistant.Recipe.DiscoveryRecipe
 import kotlinx.serialization.json.JSON
 import java.util.*
 
@@ -14,30 +16,17 @@ class SearchDiscovery(private val inputListener: OnTaskCompleted):
 
     override fun doInBackground(vararg args: Array<Ingredient>): MutableList<DiscoveryRecipe> {
         val ingredients = args[0]
-        val orderedRecipes = PriorityQueue<DiscoveryRecipe>(10) { a, b ->
-            when {
-                a.queueValue < (b.queueValue) -> 1
-                else -> -1
-            }
-        }
         val discovery = Discovery(VERSION_DIS, USERNAME_MIKE_DIS, PASSWORD_MIKE_DIS)
         discovery.endPoint = URL_MIKE_DIS
 
         val queryBuilder = QueryOptions.Builder(ENV_ID_MIKE_DIS, COL_ID_MIKE_DIS)
 
-        queryBuilder.query(buildIngredientQuery(ingredients)).count(50)
+        queryBuilder.query(buildIngredientQuery(ingredients)).count(DIS_COUNT)
 
         try {
             val queryResponse = discovery.query(queryBuilder.build()).execute()
-            for (response in queryResponse.results) {
-                val recipe = JSON.nonstrict.parse<DiscoveryRecipe>(response.toString())
-                recipe.calculatePercentAvailable(ingredients)
-                orderedRecipes.add(recipe)
-            }
-            orderedRecipes.filter {
-                it.title != "" && it.ingredientList.count() > 0
-            }
-            return orderedRecipes.toMutableList()
+
+            return this.processResponse(queryResponse, ingredients)
         } catch (exception: NotFoundException) {
             Log.d("Discovery Down", "Discovery service is not working.")
         }
@@ -49,8 +38,31 @@ class SearchDiscovery(private val inputListener: OnTaskCompleted):
         inputListener.onTaskCompleted(result)
     }
 
-    private fun buildIngredientQuery(ingredients: Array<Ingredient>): String{
+    fun processResponse(input: QueryResponse, ingredients: Array<Ingredient>): MutableList<DiscoveryRecipe>{
+
+        val orderedRecipes = PriorityQueue<DiscoveryRecipe>(10) { a, b ->
+            when {
+                a.queueValue < (b.queueValue) -> 1
+                else -> -1
+            }
+        }
+
+        for (response in input.results) {
+            val recipe = JSON.nonstrict.parse<DiscoveryRecipe>(response.toString())
+            recipe.calculatePercentAvailable(ingredients)
+            orderedRecipes.add(recipe)
+        }
+        orderedRecipes.filter {
+            it.title != "" && it.ingredientList.count() > 0
+        }
+
+        return orderedRecipes.toMutableList()
+    }
+
+    fun buildIngredientQuery(ingredients: Array<Ingredient>): String{
         return ingredients.asSequence().filter { it.name != "" }
                 .joinToString("|", "ingredientList:") { it.name }
     }
+
+
 }
