+++ app/src/main/java/com/speakeasy/watsonbarassistant/Recipe/MyRecipesTab.kt
@@ -1,4 +1,4 @@
-package com.speakeasy.watsonbarassistant
+package com.speakeasy.watsonbarassistant.Recipe
 
 import android.content.Intent
 import android.os.Bundle
@@ -9,6 +9,7 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.speakeasy.watsonbarassistant.*
 import kotlinx.android.synthetic.main.fragment_my_recipes_tab.*
 
 class MyRecipesTab : Fragment() {
