+++ app/src/main/java/com/speakeasy/watsonbarassistant/Recipe/Recipe.kt
@@ -1,0 +1,46 @@
+package com.speakeasy.watsonbarassistant.Recipe
+
+import com.speakeasy.watsonbarassistant.Ingredient
+import kotlinx.serialization.Optional
+import java.io.Serializable
+
+@kotlinx.serialization.Serializable
+
+data class DiscoveryRecipe(@Optional var queueValue: Int = 0,
+                           @Optional val title: String = "",
+                           @Optional val imageUri: String = "",
+                           @Optional val reviewCount: String = "",
+                           @Optional val description: String = "",
+                           @Optional val recipeUrl: String = "",
+                           @Optional val ingredientList: List<String> = emptyList(),
+                           @Optional val instructionList: List<String> = emptyList(),
+                           @Optional val prepTime: String = "",
+                           @Optional val cookTime: String = "",
+                           @Optional val totalTime: String = "",
+                           @Optional val imageId: String = "",
+                           @Optional val googleBestImgUrl: String = "",
+                           @Optional val googleBestImgScore: Double = 0.0): Serializable {
+
+    fun calculatePercentAvailable(userIngredients: Array<Ingredient>){
+
+        var count = 0
+        for(recipeIngredient in ingredientList) {
+            for(userIngredient in userIngredients) {
+                if(recipeIngredient.contains(userIngredient.name, ignoreCase = true)) {
+                    count++
+                    break
+                }
+            }
+        }
+
+        if(ingredientList.count() > 0) {
+            queueValue = (count * 100) / (ingredientList.count())
+        } else{
+            queueValue = 0
+        }
+    }
+
+    fun getImageName(): String {
+        return "recipe_images/GSBimg-${imageId.toFloat().toInt()}.jpg"
+    }
+}
