+++ app/src/main/res/layout/activity_recipe_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RecipeDetail">
+    tools:context=".Recipe.RecipeDetail">
 
     <LinearLayout
         android:layout_marginBottom="20dp"
