+++ app/src/main/res/layout/fragment_my_recipes_tab.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MyRecipesTab"
+    tools:context=".Recipe.MyRecipesTab"
     android:orientation="vertical">
 
     <TextView
